services/auth-service/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.10"
anyio==4.9.0 ; python_version >= "3.10"
certifi==2025.7.14 ; python_version >= "3.10"
charset-normalizer==3.4.2 ; python_version >= "3.10"
click==8.2.1 ; python_version >= "3.10"
colorama==0.4.6 ; python_version >= "3.10" and platform_system == "Windows"
deprecated==1.2.18 ; python_version >= "3.10" and python_version < "4.0"
ecdsa==0.19.1 ; python_version >= "3.10"
exceptiongroup==1.3.0 ; python_version == "3.10"
fastapi==0.116.1 ; python_version >= "3.10"
greenlet==3.2.3 ; python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.10"
h11==0.16.0 ; python_version >= "3.10"
idna==3.10 ; python_version >= "3.10"
limits==5.4.0 ; python_version >= "3.10" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.10" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.10"
psycopg2-binary==2.9.10 ; python_version >= "3.10"
pyasn1==0.6.1 ; python_version >= "3.10"
pydantic-core==2.33.2 ; python_version >= "3.10"
pydantic==2.11.7 ; python_version >= "3.10"
python-dotenv==1.1.1 ; python_version >= "3.10"
python-jose==3.5.0 ; python_version >= "3.10"
requests==2.32.4 ; python_version >= "3.10"
rsa==4.2 ; python_version >= "3.10"
six==1.17.0 ; python_version >= "3.10"
slowapi==0.1.9 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.10"
sqlalchemy==2.0.41 ; python_version >= "3.10"
starlette==0.47.2 ; python_version >= "3.10"
typing-extensions==4.14.1 ; python_version >= "3.10"
typing-inspection==0.4.1 ; python_version >= "3.10"
urllib3==2.5.0 ; python_version >= "3.10"
uvicorn==0.35.0 ; python_version >= "3.10"
wrapt==1.17.2 ; python_version >= "3.10" and python_version < "4.0"
